refactor: move all of the util function into one file
